Update to new Modality
In the device the same midi channel / numbers are used by more than one element. This would make the description of all four "Programs" messy and the use confuse. So I left to the user to edit the programs settings in the MPK MiniMKII editor to avoid the problem. Just choosing different midi channels for the different element types should do it.

 Changed midiMsgType: \noteOnOff to the new way of having the same functionality using groupType: \noteOnOffBut