making a group for all pads
Changing again to name properly pads as 'pad'.
I was trying to make the path to the elements shorter, so to access and element, i.e a pad in the hardware Bank A from the midi cc message type, one could just write elAt(\pcA, x).
The two problems with that are: there's no "pcA" anywhere else in the Modality descs, and also would not be easy to treat all the pads, or all the pads from the bank A as a group, so it is less modal.
I'm then reintroducing a longer path, but I think is for the better. I accept suggestions about naming the pads. I'm calling them \note for the midi-note-type ones and \ccm for the cc-midi-type.